enh: use iter.Map for better concurrency; fix: cmd flags
