+++ cmd/root.go
@@ -36,23 +36,19 @@ func init() {
 	logger = logging.GetLogManager()
 	rootCmd.PersistentFlags().BoolP(flagVerbose, "v", false, "Verbose output")
 	rootCmd.PersistentFlags().BoolP(flagDebug, "d", false, "Debug output")
-	dumpCmd.PersistentFlags().StringVarP(&awsProfile, flagAWSProfile, "p", "default", "AWS Profile to use")
-	dumpCmd.PersistentFlags().StringVarP(&outputDirectory, flagOutputDirectory, "o", "", "Output folder where the files will be saved (default: \".\")")
-	dumpCmd.PersistentFlags().StringVarP(&outputFormat, flagOutputFormat, "f", "zip", "Output format: ZIP or json files")
-	dumpCmd.PersistentFlags().BoolVarP(&dumpOnly, flagDumpOnly, "", false, "Flag to prevent loading data into Neo4j (default: \"false\")")
+	dumpCmd.Flags().StringVarP(&awsProfile, flagAWSProfile, "p", "default", "AWS Profile to use")
+	dumpCmd.Flags().StringVarP(&outputDirectory, flagOutputDirectory, "o", "", "Output folder where the files will be saved (default: \".\")")
+	dumpCmd.Flags().StringVarP(&outputFormat, flagOutputFormat, "f", "zip", "Output format: ZIP or json files")
+	dumpCmd.Flags().BoolVarP(&dumpOnly, flagDumpOnly, "", false, "Flag to prevent loading data into Neo4j (default: \"false\")")
+	_ = dumpCmd.MarkFlagRequired(flagAWSProfile)
 
-	assessCmd.PersistentFlags().StringVarP(&importFile, flagImportFile, "i", "", "Input ZIP file to load")
-	assessCmd.PersistentFlags().BoolVarP(&noImport, flagNoImport, "", false, "Use stored data from Neo4j without import (default)")
+	assessCmd.Flags().StringVarP(&importFile, flagImportFile, "i", "", "Input ZIP file to load")
+	assessCmd.Flags().BoolVarP(&noImport, flagNoImport, "", false, "Use stored data from Neo4j without import (default)")
+	assessCmd.MarkFlagsMutuallyExclusive(flagImportFile, flagNoImport)
 }
 
 func Execute() {
 	if err := rootCmd.Execute(); err != nil {
 		logger.Error("Error executing command", "err", err)
-	}
-}
-
-func markAsRequired(flag string) {
-	if err := rootCmd.MarkFlagRequired(flag); err != nil {
-		logger.Error("Required flags not provided", "err", err, "flag", flag)
 	}
 }
