+++ pkg/connector/services/aws/aws.go
@@ -4,12 +4,12 @@ import (
 	"context"
 	"os"
 
-	"github.com/primait/nuvola/connector/services/aws/database"
-	"github.com/primait/nuvola/connector/services/aws/ec2"
-	"github.com/primait/nuvola/connector/services/aws/iam"
-	"github.com/primait/nuvola/connector/services/aws/lambda"
-	"github.com/primait/nuvola/connector/services/aws/s3"
-	"github.com/primait/nuvola/connector/services/aws/sts"
+	"github.com/primait/nuvola/pkg/connector/services/aws/database"
+	"github.com/primait/nuvola/pkg/connector/services/aws/ec2"
+	"github.com/primait/nuvola/pkg/connector/services/aws/iam"
+	"github.com/primait/nuvola/pkg/connector/services/aws/lambda"
+	"github.com/primait/nuvola/pkg/connector/services/aws/s3"
+	"github.com/primait/nuvola/pkg/connector/services/aws/sts"
 	"github.com/primait/nuvola/pkg/io/logging"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
