+++ pkg/connector/services/aws/database/structs.go
