+++ pkg/connector/services/aws/ec2/structs.go
