+++ pkg/connector/services/aws/lambda/lambda.go
@@ -4,16 +4,15 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"sync"
 
-	"github.com/primait/nuvola/connector/services/aws/ec2"
+	"github.com/primait/nuvola/pkg/connector/services/aws/ec2"
 	"github.com/primait/nuvola/pkg/io/logging"
+	"github.com/sourcegraph/conc/iter"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/aws/transport/http"
 	"github.com/aws/aws-sdk-go-v2/service/lambda"
 	"github.com/aws/aws-sdk-go-v2/service/lambda/types"
-	"golang.org/x/sync/semaphore"
 )
 
 // aws iam list-users
@@ -30,37 +29,19 @@ func ListFunctions(cfg aws.Config) (lambdas []*Lambda) {
 }
 
 func (lc *LambdaClient) listFunctionsForRegion() (lambdas []*Lambda) {
-	var (
-		mu  = &sync.Mutex{}
-		sem = semaphore.NewWeighted(int64(20))
-		wg  sync.WaitGroup
-	)
-
 	output, err := lc.client.ListFunctions(context.TODO(), &lambda.ListFunctionsInput{})
 
 	if errors.As(err, &re) {
 		logging.HandleAWSError(re, "Lambda", "ListFunctions")
 	}
 
-	for _, lambda := range output.Functions {
-		wg.Add(1)
-		go func(lambda types.FunctionConfiguration) {
-			if err := sem.Acquire(context.Background(), 1); err != nil {
-				logging.HandleError(err, "Lambda", "listFunctionsForRegion - Acquire Semaphore")
-			}
-			defer sem.Release(1)
-			defer wg.Done()
-			defer mu.Unlock()
-			var item = &Lambda{
-				FunctionConfiguration: lambda,
-				FunctionCodeLocation:  lc.getFunctionCodeLocation(aws.ToString(lambda.FunctionName)),
-				Policy:                lc.getPolicy(aws.ToString(lambda.FunctionName)),
-			}
-			mu.Lock()
-			lambdas = append(lambdas, item)
-		}(lambda)
-	}
-	wg.Wait()
+	lambdas = iter.Map(output.Functions, func(lambda *types.FunctionConfiguration) *Lambda {
+		return &Lambda{
+			FunctionConfiguration: *lambda,
+			FunctionCodeLocation:  lc.getFunctionCodeLocation(aws.ToString(lambda.FunctionName)),
+			Policy:                lc.getPolicy(aws.ToString(lambda.FunctionName)),
+		}
+	})
 	return
 }
