+++ pkg/connector/services/aws/lambda/structs.go
