+++ pkg/connector/services/aws/s3/structs.go
