+++ pkg/connector/services/aws/structs.go
