+++ pkg/connector/services/aws/sts/sts.go
