+++ pkg/connector/services/aws/tools.go
